build: move talvos build up a level
This snapshot corresponds with changes to
https://github.com/sethp/learn-gpgpu that build the project as a part of
a larger CMake "superproject" that includes the SPIRV-Tools and
SPIRV-Headers dependencies.

It removes the docker build tooling from this project, as none of the
directories quite line up anymore to get the DWARF paths right, and
making it work reliably in both situations would require a level of CI
tooling that we don't currently have to validate that the build still
works and produces a useful output.