wip: implement experimental exec opcodes
Implement basic support for a trivial tcf program associated with the
`fill_idx` kernel that permits the spvasm program to contain all the
relevant details of how it ought to be invoked.

Still missing:

* Support for `SPECIALIZE`
* Support for preloading a buffer with arbitrary data (e.g. `BUFFER ...
  DATA ...` or `BUFFER ... BINFILE ...`)

But the former isn't required (yet) for our GPGPU playground goals, and
the latter will take some different effort to support well in any event.

See also:
https://github.com/sethp/SPIRV-Headers/commit/4f3a0d30a4e168b242ff791c611a5572b9d87fc6

It also leaves the "dispatch" work intentionally unfinished; there are
examples for which that's compelling (reductions, matrix ops, etc.), but
the example we're currently working (filling a vector with the index of
its slot) as a way to explore an information-dense UX doesn't need it.